Fix matching checkpoint index types

+++ lib/services/spotify_track_matching_service.dart
@@ -68,12 +68,14 @@ class SpotifyTrackMatchingService {
     if (tracks.isEmpty) return _snapshot(tracks, results, metadata);
 
     var nextIndex = _asInt(metadata['nextTrackIndex']) ?? results.length;
-    nextIndex = nextIndex.clamp(0, tracks.length);
+    nextIndex = nextIndex.clamp(0, tracks.length).toInt();
     if (results.length > nextIndex) {
       results.removeRange(nextIndex, results.length);
     }
 
-    final stopIndex = (nextIndex + batchSize).clamp(0, tracks.length);
+    final stopIndex = (nextIndex + batchSize)
+        .clamp(0, tracks.length)
+        .toInt();
     metadata['matchingStatus'] = 'running';
     await _checkpoint(results, metadata, nextIndex, tracks.length);
 
@@ -238,7 +240,8 @@ class SpotifyTrackMatchingService {
     Map<String, dynamic> metadata,
   ) {
     final nextIndex = (_asInt(metadata['nextTrackIndex']) ?? results.length)
-        .clamp(0, tracks.length);
+        .clamp(0, tracks.length)
+        .toInt();
     return SpotifyMatchingSnapshot(
       totalTracks: tracks.length,
       nextTrackIndex: nextIndex,
